AppData/smart_lock/requirements.txt
opencv-python
opencv-contrib-python
pandas
datetime
configparser
numpy
RPi.Gpio